requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
blas=1.0=mkl
brotli=1.1.0=hd590300_1
brotli-bin=1.1.0=hd590300_1
brotli-python=1.1.0=py310hc6cd4ac_1
bzip2=1.0.8=hd590300_5
ca-certificates=2024.2.2=hbcca054_0
certifi=2024.2.2=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
contourpy=1.2.1=py310hd41b1e2_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.12.1=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
expat=2.6.2=h59595ed_0
ffmpeg=4.3=hf484d3e_0
filelock=3.14.0=pyhd8ed1ab_0
fontconfig=2.14.2=h14ed4e7_0
fonttools=4.51.0=py310h2372a71_0
freetype=2.12.1=h267a509_2
glib=2.80.2=hf974151_0
glib-tools=2.80.2=hb6ce0ca_0
gmp=6.3.0=h59595ed_1
gmpy2=2.1.5=py310hc7909c9_1
gnutls=3.6.13=h85f3911_1
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=73.2=h59595ed_0
idna=3.7=pyhd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=pyhd8ed1ab_0
jpeg=9e=h0b41bf4_3
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=py310hd41b1e2_1
krb5=1.20.1=h81ceb04_0
lame=3.100=h166bdaf_1003
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.40=h55db66e_0
lerc=4.0.0=h27087fc_0
libblas=3.9.0=1_h86c2bf4_netlib
libbrotlicommon=1.1.0=hd590300_1
libbrotlidec=1.1.0=hd590300_1
libbrotlienc=1.1.0=hd590300_1
libcblas=3.9.0=5_h92ddd45_netlib
libclang=14.0.6=default_h7634d5b_1
libclang13=14.0.6=default_h9986a30_1
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.9.1.3=0
libcups=2.3.3=h36d4200_3
libcurand=10.3.5.147=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h0b41bf4_0
libedit=3.1.20191231=he28a2e2_2
libexpat=2.6.2=h59595ed_0
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h77fa898_7
libgfortran-ng=13.2.0=h69a702a_7
libgfortran5=13.2.0=hca663fb_7
libglib=2.80.2=hf974151_0
libgomp=13.2.0=h77fa898_7
libhwloc=2.10.0=default_h5622ce7_1001
libiconv=1.17=hd590300_2
liblapack=3.9.0=5_h92ddd45_netlib
libllvm14=14.0.6=hcd5def8_4
libnpp=11.7.4.75=0
libnsl=2.0.1=hd590300_0
libnvjpeg=11.8.0.2=0
libpng=1.6.43=h2797004_0
libpq=12.17=hdbd6064_0
libsqlite=3.45.3=h2797004_0
libstdcxx-ng=13.2.0=hc0a3c3a_7
libtiff=4.5.0=h6adf6a1_2
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.4.0=hd590300_0
libxcb=1.16=hd590300_0
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.12.7=hc051c1a_0
libzlib=1.2.13=hd590300_5
markupsafe=2.1.5=py310h2372a71_0
matplotlib=3.8.4=py310hff52083_2
matplotlib-base=3.8.4=py310hef631a5_2
mkl=2023.1.0=h213fc3f_46344
mpc=1.3.1=hfe3b2da_0
mpfr=4.2.1=h9458935_1
mpmath=1.3.0=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
mysql=5.7.20=hf484d3e_1001
ncurses=6.5=h59595ed_0
nettle=3.6=he412f7d_0
networkx=3.3=pyhd8ed1ab_1
numpy=1.26.4=py310hb13e2d6_0
openh264=2.1.1=h780b84a_0
openjpeg=2.5.0=hfec8fc6_2
openssl=3.3.0=h4ab18f5_2
packaging=24.0=pyhd8ed1ab_0
pcre2=10.43=hcad00b1_0
pillow=10.3.0=py310h5eee18b_0
pip=24.0=pyhd8ed1ab_0
pthread-stubs=0.4=h36c2ea0_1001
pyparsing=3.1.2=pyhd8ed1ab_0
pyqt=5.15.4=py310hd8f1fbe_0
pyqt5-sip=12.9.0=py310hd8f1fbe_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.9=he550d4f_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.10=4_cp310
pytorch=2.0.0=py3.10_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
qt-main=5.15.2=h53bd1ea_10
readline=8.2=h8228510_1
requests=2.32.1=pyhd8ed1ab_0
setuptools=69.5.1=pyhd8ed1ab_0
sip=6.5.1=py310h122e73d_2
six=1.16.0=pyh6c4a22f_0
sqlite=3.45.3=h2c6b66d_0
sympy=1.12=pypyh9d50eac_103
tbb=2021.12.0=h297d8ca_1
tk=8.6.13=noxft_h4845f30_101
toml=0.10.2=pyhd8ed1ab_0
torchaudio=2.0.0=py310_cu117
torchtriton=2.0.0=py310
torchvision=0.15.0=py310_cu117
tornado=6.4=py310h2372a71_0
typing_extensions=4.11.0=pyha770c72_0
tzdata=2024a=h0c530f3_0
unicodedata2=15.1.0=py310h2372a71_0
urllib3=2.2.1=pyhd8ed1ab_0
wheel=0.43.0=pyhd8ed1ab_1
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.6=h166bdaf_0
zlib=1.2.13=hd590300_5
zstd=1.5.6=ha6fb4c9_0
